chore: foundedYear 타입 변경

+++ backend/src/main/java/zipgo/brand/domain/Brand.java
@@ -34,7 +34,7 @@ public class Brand {
     private String nation;
 
     @Column
-    private Long foundedYear;
+    private Integer foundedYear;
 
     @Column(nullable = false)
     private boolean hasResearchCenter;
